refactored generate schema methods into a new file
